moves prepend.php and Bootstrap.php to the logging client
