+++ tests/unit/Logging/fakeGlobalFunctions.php
